backend/requirements.txt
Django==3.2.6
djangorestframework==3.12.4

psycopg2-binary>=2.8
requests==2.26.0
whitenoise==5.3.0


redis==3.5.3
celery==5.1.2
django-celery-beat==2.2.1
django-redis==5.0.0
django-filter==2.4.0